Beach part 3
Was able to make the sand. I need to make a conditional statement. Need to learn from the YouTube video

I am thinking of changing the colors of the sky and ocean if the mouse goes to past a certain place.

I am also thinking I could make the birds move depending where the mouse is on the x axis.